debug delit post

+++ lesson_app/views/delete_post.py
@@ -10,8 +10,9 @@ from lesson_app.models import Post
 class DeletePpost(View):
     def get(self, request, pk):
         get_post = Post.objects.get(pk=pk)
-        path = f".\media\{get_post.image}"
-        os.remove(path)
+        if get_post.image:
+            path = f".\media\{get_post.image}"
+            os.remove(path)
         get_post.delete()
 
         return redirect(reverse('profile'))
